requirements.txt
flask==2.3.3
opencv-python==4.8.1.78
numpy==1.24.3
insightface==0.7.3
albumentations==1.3.1
torch==2.0.1
torchvision==0.15.2
python-dotenv==1.0.0
pillow==10.0.0
requests==2.31.0
werkzeug==2.3.7
onnxruntime==1.16.3
onnx==1.14.1
scipy==1.11.4
scikit-learn==1.3.2
tqdm==4.66.1